Add server-side input validation and show error messages to user
Fix a bug in JsonDataManager.UpdateProject that caused NotFoundException
 to be thrown if updated project DID exist.
Fix a typo in JsonDataManager.FindReportsByProject method name.
Fix a bug in Project/Show view that caused date not to be sent in route
 data while updating accepted time.

+++ lab1/Controllers/ReportEntryController.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using AutoMapper;
@@ -7,7 +6,9 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.Extensions.Logging;
 using TRS.Controllers.Attributes;
+using TRS.Controllers.Constants;
 using TRS.DataManager;
+using TRS.DataManager.Exceptions;
 using TRS.Extensions;
 using TRS.Models.DomainModels;
 using TRS.Models.ViewModels;
@@ -28,13 +29,33 @@ namespace TRS.Controllers
         public IActionResult Show(int id)
         {
             var reportEntry = DataManager.FindReportEntryByDayAndIndex(LoggedInUser.Name, RequestedDate, id);
+            if (reportEntry == null)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportEntryNotFoundMessage(RequestedDate, id);
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
             return View(Mapper.Map<ReportEntryModel>(reportEntry));
         }
 
         public IActionResult Edit(int id)
         {
+            if (DataManager.FindReportByUserAndMonth(LoggedInUser.Name, RequestedDate).Frozen)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportFrozenMessage(RequestedDate.ToMonthString());
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
             var reportEntry = DataManager.FindReportEntryByDayAndIndex(LoggedInUser.Name, RequestedDate, id);
+            if (reportEntry == null)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportEntryNotFoundMessage(RequestedDate, id);
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
             var project = DataManager.FindProjectByCode(reportEntry.Code);
+            if (project == null)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetProjectNotFoundMessage(reportEntry.Code);
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
             var categoryCodes = new List<SelectListItem> { new("nieokreślony", "") }.Concat(
                 project.Subactivities.Select(y => new SelectListItem(y.Code, y.Code))).ToList();
             var model = new ReportEntryForEditingModel
@@ -48,7 +69,17 @@ namespace TRS.Controllers
         [HttpPost]
         public IActionResult Edit(int id, ReportEntryUpdateModel reportEntryUpdate)
         {
+            if (DataManager.FindReportByUserAndMonth(LoggedInUser.Name, RequestedDate).Frozen)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportFrozenMessage(RequestedDate.ToMonthString());
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
             var reportEntry = DataManager.FindReportEntryByDayAndIndex(LoggedInUser.Name, RequestedDate, id);
+            if (reportEntry == null)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportEntryNotFoundMessage(RequestedDate, id);
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
             reportEntry.Subcode = reportEntryUpdate.Subcode;
             reportEntry.Time = reportEntryUpdate.Time;
             reportEntry.Description = reportEntryUpdate.Description;
@@ -58,6 +89,11 @@ namespace TRS.Controllers
 
         public IActionResult Add()
         {
+            if (DataManager.FindReportByUserAndMonth(LoggedInUser.Name, RequestedDate).Frozen)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportFrozenMessage(RequestedDate.ToMonthString());
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
             var availableProjects = DataManager.GetAllProjects().Where(x => x.Active).ToHashSet();
             var projectCodes = availableProjects.Select(x => new SelectListItem(x.Name, x.Code)).ToList();
             var categoryCodes = availableProjects.ToDictionary(x => x.Code,
@@ -76,6 +112,22 @@ namespace TRS.Controllers
         [HttpPost]
         public IActionResult Add(ReportEntryModel reportEntry)
         {
+            if (DataManager.FindReportByUserAndMonth(LoggedInUser.Name, reportEntry.Date).Frozen)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportFrozenMessage(reportEntry.Date.ToMonthString());
+                return RedirectToAction("Index", "Home", new { Date = reportEntry.Date.ToDateString() });
+            }
+            var project = DataManager.FindProjectByCode(reportEntry.Code);
+            if (project == null)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetProjectNotFoundMessage(reportEntry.Code);
+                return RedirectToAction("Add", new { Date = reportEntry.Date.ToDateString() });
+            }
+            if (!project.Active)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetProjectNoLongerActiveMessage(reportEntry.Code);
+                return RedirectToAction("Add", new { Date = reportEntry.Date.ToDateString() });
+            }
             DataManager.AddReportEntry(LoggedInUser.Name, Mapper.Map<ReportEntry>(reportEntry));
             return RedirectToAction("Index", "Home", new { Date = reportEntry.Date.ToDateString() });
         }
@@ -83,7 +135,20 @@ namespace TRS.Controllers
         [HttpPost]
         public IActionResult Delete(int id)
         {
-            DataManager.DeleteReportEntry(LoggedInUser.Name, RequestedDate, id);
+            if (DataManager.FindReportByUserAndMonth(LoggedInUser.Name, RequestedDate).Frozen)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportFrozenMessage(RequestedDate.ToMonthString());
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
+            try
+            {
+                DataManager.DeleteReportEntry(LoggedInUser.Name, RequestedDate, id);
+            }
+            catch (NotFoundException)
+            {
+                TempData[ErrorTempDataKey] = ErrorMessages.GetReportEntryNotFoundMessage(RequestedDate, id);
+                return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
+            }
             return RedirectToAction("Index", "Home", new { Date = RequestedDate.ToDateString() });
         }
